+++ src/app/(markdown)/layout.tsx
@@ -7,17 +7,19 @@ import ArrowLink from '@/components/links/ArrowLink';
 const LayoutPage = ({ children }: { children: React.ReactNode }) => {
   const pathName = usePathname();
   const arrRoute = pathName.split('/');
-  const baseLabel = arrRoute.length > 2 ? arrRoute[1] : 'Home';
+  const baseLabel = arrRoute[1];
 
   return (
     <section className='min-h-screen py-24'>
-      <ArrowLink
-        direction='left'
-        className='mb-10 mt-2'
-        href={arrRoute.length > 2 ? '/' + arrRoute[1] : '/'}
-      >
-        Back to <span className='capitalize'>{baseLabel}</span>
-      </ArrowLink>
+      {arrRoute.length > 2 && (
+        <ArrowLink
+          direction='left'
+          className='mb-10 mt-2'
+          href={'/' + arrRoute[1]}
+        >
+          Back to <span className='capitalize'>{baseLabel}</span>
+        </ArrowLink>
+      )}
       {children}
     </section>
   );