+++ src/components/form/InputSearch.tsx
@@ -1,39 +1,62 @@
-'use client';
-
 import * as React from 'react';
 
 import { cn } from '@/lib/utils';
 
-import StyledInput from '@/components/form/Input';
+export default function InputSearch({
+  className,
+  ...rest
+}: React.ComponentPropsWithoutRef<'input'>) {
+  const inputRef = React.useRef<HTMLInputElement>(null);
 
-type InputSearchProps = React.HTMLAttributes<HTMLFormElement>;
+  const keyDownHandler = (event: KeyboardEvent) => {
+    if (event.ctrlKey && event.key === 'k') {
+      event.preventDefault();
+      inputRef.current?.focus();
+    }
+  };
 
-export function InputSearch({ className, ...props }: InputSearchProps) {
-  function onSubmit(event: React.SyntheticEvent) {
-    event.preventDefault();
+  React.useEffect(() => {
+    window.addEventListener('keydown', keyDownHandler);
 
-    alert('Still Under Construction');
-
-    // return toast({
-    //   title: 'Not implemented',
-    //   description: "We're still working on the search.",
-    // });
-  }
+    return () => {
+      window.removeEventListener('keydown', keyDownHandler);
+    };
+  }, []);
 
   return (
-    <form
-      onSubmit={onSubmit}
-      className={cn('relative w-full', className)}
-      {...props}
-    >
-      <StyledInput
-        type='search'
-        placeholder='Search blog, project, docs...'
-        className='bg-base-100/50 h-8 w-full sm:w-64 sm:pr-12'
-      />
-      <kbd className='border-base-content/70 pointer-events-none absolute right-1.5 top-1.5 hidden h-5 select-none items-center justify-center gap-1 rounded border px-1.5 font-mono text-[10px] font-medium sm:flex'>
-        <span className='text-xs'>⌘</span>K
-      </kbd>
-    </form>
+    <>
+      <div className={cn('relative')}>
+        <input
+          className={cn(
+            className,
+            'pr-24',
+            'bg-base-100 h-10 w-full rounded-md',
+            'border border-gray-300 dark:border-gray-600',
+            'focus:border-primary dark:focus:border-primary focus:outline-none focus:ring-0',
+            'text-sm'
+          )}
+          ref={inputRef}
+          {...rest}
+        />
+        <kbd
+          className={cn(
+            'absolute right-1.5 top-[.6rem]',
+            'h-5 rounded px-1.5',
+            'border border-gray-300 dark:border-gray-600',
+            'pointer-events-none select-none',
+            'font-mono text-[10px] font-medium'
+          )}
+        >
+          <span
+            className={cn(
+              'flex items-center justify-center',
+              'text-xs leading-relaxed'
+            )}
+          >
+            Ctrl + K
+          </span>
+        </kbd>
+      </div>
+    </>
   );
 }