+++ src/components/form/StyledInput.tsx
@@ -7,14 +7,19 @@ export default function StyledInput({
   ...rest
 }: React.ComponentPropsWithoutRef<'input'>) {
   return (
-    <input
-      className={cn(
-        className,
-        'bg-base-100 h-10 w-full rounded-md',
-        'border border-gray-300 dark:border-gray-600',
-        'focus:border-primary dark:focus:border-primary focus:outline-none focus:ring-0'
-      )}
-      {...rest}
-    />
+    <>
+      <div className={cn('relative')}>
+        <input
+          className={cn(
+            className,
+            'bg-base-100 h-10 w-full rounded-md',
+            'border border-gray-300 dark:border-gray-600',
+            'focus:border-primary dark:focus:border-primary focus:outline-none focus:ring-0',
+            'text-sm'
+          )}
+          {...rest}
+        />
+      </div>
+    </>
   );
 }