+++ src/components/layout/MenuMobile.tsx
@@ -3,16 +3,21 @@ import { AnimatePresence, motion } from 'framer-motion';
 import { useTheme } from 'next-themes';
 import React, { useRef, useState } from 'react';
 
+import { cn } from '@/lib/utils';
 import { useClickOutside } from '@/hooks/useClickOutside';
 
-import ReactScroll from '@/components/links/ReactScroll';
+import UnstyledLink from '@/components/links/UnstyledLink';
 import CloseIcon from '@/components/settings/CloseIcon';
 import HamburgerIcon from '@/components/settings/HamburgerIcon';
 import ThemeChanger from '@/components/settings/ThemeChanger';
 
 import { NavigationData } from '@/constant/config';
 
-const ModalSetting = () => {
+type MenuMobileProps = {
+  activePage: string;
+};
+
+const MenuMobile = ({ activePage }: MenuMobileProps) => {
   const [isOpen, setIsOpen] = useState(false);
   const { theme, setTheme } = useTheme();
 
@@ -32,9 +37,9 @@ const ModalSetting = () => {
       ref={boxRef}
     >
       {/* TRIGGER BUTTON */}
-      <label htmlFor='toggle-modal' className='swap swap-rotate'>
+      <label htmlFor='toggle-menu-mobile' className='swap swap-rotate'>
         <input
-          id='toggle-modal'
+          id='toggle-menu-mobile'
           type='checkbox'
           className='hidden'
           checked={isOpen}
@@ -57,7 +62,7 @@ const ModalSetting = () => {
             animate={{ opacity: 1, y: 0 }}
             exit={{ opacity: 0, y: -10 }}
             transition={{ duration: 0.3 }}
-            className='absolute right-0 top-8 z-[100]'
+            className='absolute right-0 top-8 z-[999]'
           >
             <ul
               tabIndex={0}
@@ -65,13 +70,15 @@ const ModalSetting = () => {
             >
               {NavigationData.map((nav) => (
                 <li key={nav.id}>
-                  <ReactScroll
-                    to={nav.value}
+                  <UnstyledLink
                     href={nav.link}
-                    className='w-full p-2'
+                    className={cn(
+                      { active: nav.link === activePage },
+                      'w-full p-2'
+                    )}
                   >
                     {nav.value}
-                  </ReactScroll>
+                  </UnstyledLink>
                 </li>
               ))}
 
@@ -88,4 +95,4 @@ const ModalSetting = () => {
   );
 };
 
-export default ModalSetting;
+export default MenuMobile;